requirements.txt
uv
strands-agents
strands-agents-tools
mcp
icecream
pandas
numpy
langchain-community
awscurl